Refactor live mode reload hook out into `useLiveMode`
